Fix this bug: Pressing the [Page Up] key correctly displays the string "PageUp," but pressing the [Page Down] key incorrectly displays the string "Next."
This happens because Windows sends the value Keys.Next when the [Page Down] key is pressed, and the value is just converted to string before display.

To fix this, add an entry to ReplaceKey.cs:Replacements so that ReplaceKey.cs:Sanitise() will return "PageDown" when given the value Keys.Next.

System.Windows.Forms.Keys assigns the value 33 to both Prior and PageUp, and assigns the value 34 to both Next and PageDown.

It may be that Windows correctly returns Keys.PageUp instead of Keys.Prior because "PageUp" < "Prior," and incorrectly returns Keys.Next instead of Keys.PageDown because "Next" < "PageDown."

In any case, replacing Keys.Next with "PageDown" makes the output more consistent.